+++ tests/testthat/test-cache.R
@@ -1,13 +1,13 @@
 context ("cache")
 
-testthat::skip_on_cran ()
-
 # The cache-off tests fail on windows for some reason, both on CRAN and on
 # GitHub runners.
 testthat::skip_on_os ("windows")
 
 test_all <- (identical (Sys.getenv ("MPADGE_LOCAL"), "true") |
     identical (Sys.getenv ("GITHUB_WORKFLOW"), "test-coverage"))
+
+testthat::skip_if (!test_all)
 
 source ("../sc-conversion-fns.R")
 
@@ -50,24 +50,43 @@ test_that ("cache on", {
         turn_penalty = TRUE, left_side = TRUE
     )
     # grapht has extra compound edges for turning angles:
-    expect_true (nrow (grapht) > nrow (graph))
+    expect_equal (nrow (grapht), nrow (graph))
     grapht_c <- dodgr_contract_graph (grapht)
-    expect_true (nrow (grapht_c) > nrow (graph_c))
-    expect_silent (graphtf <- dodgr_flows_aggregate (grapht_c,
-        from = pts,
-        to = pts,
-        flow = fmat,
-        contract = FALSE
-    ))
-    expect_silent (graphtf <- dodgr_uncontract_graph (graphtf))
+    expect_equal (nrow (grapht_c), nrow (graph_c))
+    expect_warning (
+        graphtf <- dodgr_flows_aggregate (grapht_c,
+            from = pts,
+            to = pts,
+            flow = fmat,
+            contract = FALSE
+        ),
+        "graphs with turn penalties should be submitted in full, not contracted form"
+    )
+    expect_silent (
+        graphtf <- dodgr_flows_aggregate (grapht,
+            from = pts,
+            to = pts,
+            flow = fmat,
+            contract = FALSE
+        )
+    )
     expect_silent (graphtf <- merge_directed_graph (graphtf))
 
-    expect_silent (graphtf <-
-        dodgr_flows_disperse (grapht_c,
+    expect_warning (
+        graphtf <- dodgr_flows_disperse (
+            grapht_c,
             from = pts,
             dens = rep (1, n)
-        ))
-
+        ),
+        "graphs with turn penalties should be submitted in full, not contracted form"
+    )
+    expect_silent (
+        graphtf <- dodgr_flows_disperse (
+            grapht,
+            from = pts,
+            dens = rep (1, n)
+        )
+    )
 })
 
 test_that ("cache off", {
@@ -106,21 +125,35 @@ test_that ("cache off", {
         left_side = TRUE
     ))
     expect_silent (grapht_c <- dodgr_contract_graph (grapht))
-    expect_silent (graphtf <- dodgr_flows_aggregate (grapht_c,
+    expect_warning (
+        graphtf <- dodgr_flows_aggregate (
+            grapht_c,
+            from = pts,
+            to = pts,
+            flow = fmat,
+            contract = FALSE
+        ),
+        "graphs with turn penalties should be submitted in full, not contracted form"
+    )
+    expect_silent (graphtf <- dodgr_flows_aggregate (grapht,
         from = pts,
         to = pts,
         flow = fmat,
         contract = FALSE
     ))
-    expect_silent (graphtf <- dodgr_uncontract_graph (graphtf))
     expect_silent (graphtf <- merge_directed_graph (graphtf))
 
-    expect_silent (graphtf <-
-        dodgr_flows_disperse (grapht_c,
+    expect_warning (
+        graphtf <- dodgr_flows_disperse (
+            grapht_c,
             from = pts,
             dens = rep (1, n)
-        ))
+        ),
+        "graphs with turn penalties should be submitted in full, not contracted form"
+    )
+    expect_silent (
+        graphtf <- dodgr_flows_disperse (grapht, from = pts, dens = rep (1, n))
+    )
 
     expect_silent (dodgr_cache_on ())
-
 })